add line drawing

+++ SGUI_string.c
@@ -8,7 +8,7 @@ static char strBuf[64] = {0};
 void SGUI_printString(char* string,
                       unsigned short x, unsigned short y,
                       unsigned short size,
-                      unsigned short backgroundColor, unsigned short textColor)
+                      signed long backgroundColor, unsigned short textColor)
 {
   char *nlL = string;
   char *nlR = 0;
